refactor(wan): colocate video VAE codecs
