+++ unirl/models/wan21/__init__.py
@@ -28,8 +28,7 @@ from unirl.models.wan21.diffusion import (
 from unirl.models.wan21.image_encode import WAN21ImageLatentEncodeStage
 from unirl.models.wan21.pipeline import WAN21Pipeline
 from unirl.models.wan21.text_embed import WAN21TextEmbedStage
-from unirl.models.wan21.vae import WAN21VAEDecodeStage
-from unirl.models.wan21.video_encode import WANVideoLatentEncodeStage
+from unirl.models.wan21.vae import WAN21VAEDecodeStage, WANVideoLatentEncodeStage
 
 __all__ = [
     "WAN21Bundle",
